Seperated bmp export, UI cleanup, render start cleanup, mesh subdivision surfaces

+++ src/common/cpp/dummyRenderer/dummyRenderer.cpp
@@ -45,7 +45,7 @@ void DummyRenderer::doProgressiveRendering()
 	int runs = 6;
 	for( int run = runs; run > 0; run--)
 	{		
-		int stepSizeX = pow(2.0f, run);
+		int stepSizeX = (int)pow(2.0f, run);
 		int stepSizeY = stepSizeX;
 
 	}
